Implements the L2-Norm calculation of the document class vector.
Used to normalize the values of the vector to the common norm. 
When this (or any other possible) norm is used it must be used over the
whole process!

The usage of the norm does not change anything within the classification
result. So it may be "only" a performance boost.